requirements.txt
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.1
fonttools==4.59.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
setuptools==80.9.0
six==1.17.0
Werkzeug==3.1.3
wheel==0.45.1